[Edit] Edit index.js, user.js
Edit routing GET '/logout' in index.js
Edit routing DELETE '/user' in user.js and try-catch syntax

+++ server/routes/user.js
@@ -10,23 +10,17 @@ const wrap = require('express-async-wrap');
 
 // add user
 router.post('/', wrap(async (req, res) => {
-  try {
-    const create = await models.user.create(req.body);
-    if (create) {
-      res.send({
-        result: true
-      });
-    }
-  } catch (e) {
+  const create = await models.user.create(req.body);
+  if (create) {
     res.send({
-      result: false
+      result: true
     });
   }
 }));
 
 // 유저 삭제
 router.delete('/', wrap(async (req, res) => {
-  try {
+  if (req.session.user.auth === 1) {
     const dest = await models.user.destroy({
       where: { uid: req.body.uid }
     });
@@ -34,12 +28,17 @@ router.delete('/', wrap(async (req, res) => {
       res.send({
         result: true
       });
-    } else {
+    }
+  } else if (req.session.user.uid === req.body.uid) {
+    const dest = await models.user.destroy({
+      where: { uid: req.body.uid }
+    });
+    if (dest) {
       res.send({
-        result: false
+        result: true
       });
     }
-  } catch (e) {
+  } else {
     res.send({
       result: false
     });
@@ -49,21 +48,15 @@ router.delete('/', wrap(async (req, res) => {
 
 // 유저  정보 수정
 router.put('/', wrap(async (req, res) => {
-  try {
-    const update = await models.user.update(req.body, {
-      where: {
-        uid: req.session.user.uid
-      }
-    });
-
-    if (update) {
-      res.send({
-        result: true
-      });
+  const update = await models.user.update(req.body, {
+    where: {
+      uid: req.session.user.uid
     }
-  } catch (e) {
+  });
+
+  if (update) {
     res.send({
-      result: false
+      result: true
     });
   }
 }));
@@ -71,25 +64,15 @@ router.put('/', wrap(async (req, res) => {
 
 // 유저 정보 읽기
 router.get('/:uid', wrap(async (req, res) => {
-  try {
-    const inform = await models.user.findOne({
-      where: {
-        uid: req.params.uid
-      }
-    });
-    await delete inform.dataValues.pw;
-
-    if (inform) {
-      res.send(inform);
-    } else {
-      res.send({
-        result: false
-      });
+  const inform = await models.user.findOne({
+    where: {
+      uid: req.params.uid
     }
-  } catch (e) {
-    res.send({
-      result: false
-    });
+  });
+  await delete inform.dataValues.pw;
+
+  if (inform) {
+    res.send(inform);
   }
 }));
